refactor!: move TileGroup into separate module
This makes `lib.rs` smaller and more organized.
