+++ src/tile_group.rs
@@ -1,0 +1,56 @@
+use crate::hand::error::HandErr;
+use crate::{GroupType, Suit};
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct TileGroup {
+    pub value: String,
+    pub suit: Suit,
+    pub isopen: bool,
+    pub group_type: GroupType,
+    pub isterminal: bool,
+}
+
+impl TileGroup {
+    pub fn new(group: String) -> Result<Self, HandErr> {
+        let isopen = group.chars().last().unwrap().to_string() == "o";
+        let value = group.chars().nth(0).unwrap().to_string();
+
+        let suit = if !isopen {
+            group.chars().last().unwrap().to_string()
+        } else {
+            group.chars().nth(group.len() - 2).unwrap().to_string()
+        };
+        let suit = Suit::suit_from_string(suit)?;
+
+        let group_type = GroupType::group_type_from_string(group.to_string())?;
+
+        let mut isterminal = false;
+        if group_type == GroupType::Sequence {
+            if value == "1" || value == "7" {
+                isterminal = true;
+            }
+        } else if value == "1" || value == "9" {
+            isterminal = true;
+        }
+
+        let tile = Self {
+            value,
+            suit,
+            isopen,
+            group_type,
+            isterminal,
+        };
+
+        Ok(tile)
+    }
+
+    /// Check if the group is an honor.
+    pub fn is_honor(&self) -> bool {
+        matches!(self.suit, Suit::Wind | Suit::Dragon)
+    }
+
+    /// Parse the group value into an integer.
+    pub fn parse_u8(&self) -> Result<u8, std::num::ParseIntError> {
+        self.value.parse()
+    }
+}
